+++ django_echarts/__init__.py
@@ -1,0 +1,6 @@
+"""
+A django app for Echarts integration with pyecharts as chart builder.
+"""
+
+__version__ = '0.1.4'
+__author__ = 'kinegratii'